Remove AX_* macros from cfgaux/, use system versions instead.
Autoconf-archive package is now required for bootstrapping. Should not be an
issue since users usually build from bootstrapped tarballs.

To protect HttpPaxxer developers, we now also fail bootstrapping if macros
from Autoconf Macros Archive are found in configure. Autoconf is supposed to
expand Archive macros when they are installed, but it does not fail if it does
not know about those macros, and ./configure later fails in various, often
misleading, ways. We now fail at bootstrapping time with an appropriate error
message.

The main benefit is that we do not need to manually update the bundled
macros.