Added initial implementation of HPACK header field parser and packer.
Needs a lot of work. Here are some of the biggest missing pieces:
* HeaderField type is hard-coded and not user-configurable.
* Headers storage types are hard-coded and not user-configurable.
* No integration with integer paxxer.
* Testing code duplication. Tests are not a part of the library test suite.

The low-level storage classes for paxxers are difficult to get right because
they need to satisfy several design constraints:

1. The parser should not maintain information it does not need, such as info
   related to searching for fields by their names.

2. The packer should not duplicate parser storage code, including the code
   that maintains limited HPACK table size. However, the size-limiting code
   (shared by both classes) knows which specific fields are being evicted and,
   hence, needs to update the field map (specific to the packer class).

3. Library users will want to store header fields without copying, using
   user-specific types rather than std::strings. They may also supply their
   own types implementing header search and queuing.

This implementation attempts to use templates to solve the above problems.
This works quite well for Parser and the Static Table. However, the Packer
storage looks very complex. Moreover, the solution for #3 is not finalized
yet, so it is not clear whether the current version is sufficiently flexible
to support it. We may need to redesign.